==> 02-quality.qmd <==
---
title: "Quality checks"
---

Because we are extracting this data from PDF's I used this file to troubleshoot and make sure all this is good.

## Fixed issues

- In `others` there were initially some missing players. I ended up piecing everything together in [Cleaning other](01-cleaning-other.qmd).
- In the [Cleaning roster](01-cleaning-roster.qmd) notebook the `roster_type` == "SUPPLEMENTAL SPOT 31" was changed to just "SUPPLEMENTAL SPOT" for consistency. The roster has the "31" in all cases but seems odd to keep here. I can change later if needed.

There aren't any unknown issues anymore, but I'm keeping this notebook around for now.

## Setup

```{r}
#| label: setup
#| message: false
#| warning: false

library(tidyverse)
library(janitor)
```


## Import the rosters

```{r}
rosters <- read_rds("data-processed/rosters.rds")

rosters |> glimpse()
```

Do I have all the teams?

```{r}
rosters |> 
  count(club_short)
```

There are [29 teams](https://en.wikipedia.org/wiki/Major_League_Soccer) in the MLS as of May 2, 2024.

Let's spot check some teams.

```{r}
rosters |> filter(club_short == "PHI")
```

## Others

We look at the others file here.

```{r}
others <- read_rds("data-processed/others.rds")

others |> glimpse()
```

### Missing teams

At one point we were missing teams. Here we make sure there are 29.

```{r}
others |> 
  count(club_short)
```

### Checking season-ending list

I noticed this because we were missing injured players. Here I check for them.

```{r}
others |> 
  filter(player_type == "SEASON-ENDING INJURY LIST")
```

### Last check of all the data

```{r}
others |> 
  arrange(club_short)
```

## Can I join these lists somehow?

It may be possible, but not not easily. If I try a simple join, I get duplicates if a player is on more than one "others" list.

```{r}
rosters |> 
  left_join(others, by = join_by(club_short, name))
```

### Possible future solution

I might be able to do this one designation at a time

- Start with the roster
- Filter others to a specific `player_type` designation
- left_join to bring over the player_type
- rename that new column to be the specific player_type
- repeat for the other designations

Let's try this ... ah there is an issue I need to address.

```{r}
others |> 
  # select(!player_type) |> 
  # distinct() |> 
  # count(club_short,name, sort = T)
  filter(name == "Hernán López")
  
```
